snapshot_tests snapshot generator for eosio-1.8
Temporary modification of snapshot_tests so that they produce new
snapshot files (not gzipped). This is needed for forks that add new
intrinsics into the base eosio code, such as WAX. This branch is not
intended for any other development, and is for one-time use.

This branch is only usable with eosio-1.8 in order to produce v2
snapshots.